refactor: extract legacy entity-dict adapter from codegen
Continue breaking up codegen.py (still ~3k lines). Move
_program_from_legacy_mapping -- the ~160-line adapter that normalizes the
historical dict-shaped codegen input into a typed AstProgram -- into a new
codegen_legacy module. It is a pure input normalizer with no dependency on the
generator's internals, so it is a clean seam.

codegen.py re-imports it under the same private name, so the single call site
is unchanged; three now-unused AST imports and the _parse_bind_route import are
dropped. Emitted IR is byte-for-byte identical (tests/test_golden_ir.py passes
unchanged); ruff and mypy are clean.

codegen.py: 3007 -> 2845 lines.
